+++ src/rtos.c
@@ -4,6 +4,7 @@
  ******************************************************************************/
 
 #include "rtos.h"
+#include "lcd.h"
 #include "op_mode.h"
 #include "servo.h"
 #include "usart.h"
@@ -14,14 +15,17 @@
 #define TASK_DELAY_MS__TASK_OP_MODE_MGMT                50
 #define TASK_DELAY_MS__TASK_LED_CTRL                    50
 #define TASK_DELAY_MS__TASK_SERVO_MOTOR_CTRL            100
+#define TASK_DELAY_MS__TASK_LCD_CTRL                    50
 /*===== Task Priorities =====*/
 #define TASK_PRIORITY__TASK_DEFAULT                     1
 #define TASK_PRIORITY__TASK_NUCLEO_COM_PORT_IF          4
 #define TASK_PRIORITY__TASK_OP_MODE_MGMT                3
 #define TASK_PRIORITY__TASK_LED_CTRL                    2
-#define TASK_PRIORITY__TASK_SERVO_MOTOR_CTRL            5
+#define TASK_PRIORITY__TASK_SERVO_MOTOR_CTRL            6
+#define TASK_PRIORITY__TASK_LCD_CTRL                    5
 /*===== Task Stack Sizes =====*/
 #define TASK_STACK_SIZE__TASK_NUCLEO_COM_PORT_IF        (configMINIMAL_STACK_SIZE*2)
+#define TASK_STACK_SIZE__TASK_LCD_CTRL                  (configMINIMAL_STACK_SIZE*2)
 /*===== Task Handles =====*/
 static TaskHandle_t task_handle_op_mode_mgmt = NULL;
 static TaskHandle_t task_handle_led_ctrl = NULL;
@@ -33,9 +37,10 @@ static void task_nucleo_com_port_if(void *params __attribute__((unused)));
 static void task_op_mode_mgmt(void *params __attribute__((unused)));
 static void task_led_ctrl(void *params __attribute__((unused)));
 static void task_servo_motor_ctrl(void *params __attribute__((unused)));
+static void task_lcd_ctrl(void *params __attribute__((unused)));
 /*===== Other Private Functions =====*/
 static void tasks_init(void);
-static void tx_op_mode(void);
+static void tx_op_mode_to_com_port(void);
 
 /*============================================================================*/
 /*===== Public Functions =====================================================*/
@@ -84,7 +89,7 @@ static void task_nucleo_com_port_if(void *params __attribute__((unused)))
     while (1)
     {
         /* Transmit operational mode. */
-        tx_op_mode();
+        tx_op_mode_to_com_port();
 
         /* Block (delay). */
         freertos_wrapper_task_delay_ms(TASK_DELAY_MS__TASK_NUCLEO_COM_PORT_IF);
@@ -156,13 +161,57 @@ static void task_servo_motor_ctrl(void *params __attribute__((unused)))
 {
     servo_set_signal(true);
 
+    /* Idle for 5 seconds before starting. */
+    freertos_wrapper_task_delay_ms(5000);
+
     /* Task. */
     while (1)
     {
         // @todo: close the loop / implement controller / implement state-machine
 
+        /* Test feature. */
+        servo_test_oscillate(SERVO_POSITION_MIN_DEG_UINT, SERVO_POSITION_MAX_DEG_UINT, false);
+
         /* Block (delay). */
         freertos_wrapper_task_delay_ms(TASK_DELAY_MS__TASK_SERVO_MOTOR_CTRL);
+    }
+}
+
+/**
+ * @brief  RTOS task ---
+ *         LCD control.
+ * @param  params: Unused.
+ * @retval None.
+ */
+static void task_lcd_ctrl(void *params __attribute__((unused)))
+{
+    char data[LCD_MAX_DIGITS+1] = {0}; /* +1 for '\0'. */
+
+    /* Task. */
+    while (1)
+    {
+        /* Page 1, Line 1. */
+        memset(data, 0, LCD_MAX_DIGITS);
+        sprintf(data, "POS (DEG): %d", servo_get_angle_expected());
+        lcd_write_line(LCD_LINE_NUM_1, (uint8_t*)data, strlen(data));
+
+        /* Page 1, Line 2. */
+        memset(data, 0, LCD_MAX_DIGITS);
+        switch (op_mode_get())
+        {
+            case OP_MODE__UNKNOWN:           sprintf(data, "UNKNOWN");       break;
+            case OP_MODE__IDLE:              sprintf(data, "IDLE");          break;
+            case OP_MODE__MOTOR_RUNNING:     sprintf(data, "MOTOR RUNNING"); break;
+            case OP_MODE__ERROR_MOTOR:       sprintf(data, "ERROR (MOTOR)"); break;
+            case OP_MODE__ERROR_LCD:         sprintf(data, "ERROR (LCD)");   break;
+            case OP_MODE__ERROR_PERIPHERALS: sprintf(data, "ERROR (OTHER)"); break;
+            case OP_MODE__ERROR_FW_FAULT:    sprintf(data, "ERROR (FW)");    break;
+            default:                                                         break;
+        }
+        lcd_write_line(LCD_LINE_NUM_2, (uint8_t*)data, strlen(data));
+
+        /* Block (delay). */
+        freertos_wrapper_task_delay_ms(TASK_DELAY_MS__TASK_LCD_CTRL);
     }
 }
 
@@ -204,14 +253,21 @@ static void tasks_init(void)
                                  (void *)0,
                                  TASK_PRIORITY__TASK_SERVO_MOTOR_CTRL,
                                  0);
+    freertos_wrapper_task_create(task_lcd_ctrl,
+                                 "task_lcd_ctrl",
+                                 TASK_STACK_SIZE__TASK_LCD_CTRL,
+                                 (void *)0,
+                                 TASK_PRIORITY__TASK_LCD_CTRL,
+                                 0);
 }
 
 /**
  * @brief  Construct and transmit a message via the Nucleo COM port interface:
  *             - Message: operational mode.
+ *             - Message: servo motor shaft position (angle in degrees).
  * @retval None.
  */
-static void tx_op_mode(void)
+static void tx_op_mode_to_com_port(void)
 {
     #define TX_BUFF_MAX 100
     char data[TX_BUFF_MAX] = {0};
@@ -254,7 +310,12 @@ static void tx_op_mode(void)
         error_handler();
     }
 
-    /* Transmit. */
+    /* Transmit operational mode. */
+    usart_tx(handle, (uint8_t *)data, sizeof(data), 1000);
+
+    /* Transmit angle (expected). */
+    memset(data, 0, TX_BUFF_MAX);
+    sprintf(data, "Position (degrees): %d\r\n", servo_get_angle_expected());
     usart_tx(handle, (uint8_t *)data, sizeof(data), 1000);
 }